Entire project so far
Biggest changes are to the golfMain, CreateUser2Controller, LoginController, removal of CreateUser controller & fxml, and also some tweaks to FXMLs. Still planning on changing up the user creation a bit - for the Team field, it should provide a combobox of preprovided teams taken from the database. This makes sure that players on the same team have the same string for this field. Need to figure out how to populate the player table with information from the selected user. Not sure if we're keeping TreeTableView or just using a regular table.